CodeAlpha_HandwrittenCharacterRecognition/requirements.txt
tensorflow
numpy
pillow
scikit-learn
tensorflow_datasets
